planning: move util function GetSignal() down to scenario namespace since it has dependency on scenario context
