+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/BUILD
@@ -25,6 +25,7 @@ cc_library(
         "//modules/common/util:factory",
         "//modules/planning/proto:planning_proto",
         "//modules/planning/scenarios:scenario",
+        "//modules/planning/scenarios/util:scenario_util_lib",
         "//modules/planning/tasks/deciders:decider",
         "@eigen",
     ],
