+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_stop.cc
@@ -27,6 +27,7 @@
 #include "modules/planning/common/frame.h"
 #include "modules/planning/common/planning_context.h"
 #include "modules/planning/common/util/util.h"
+#include "modules/planning/scenarios/util/util.h"
 
 namespace apollo {
 namespace planning {
@@ -56,8 +57,8 @@ Stage::StageStatus TrafficLightUnprotectedRightTurnStageStop::Process(
   for (const auto& traffic_light_overlap :
        PlanningContext::GetScenarioInfo()->current_traffic_light_overlaps) {
     // check if the traffic_light is still along reference_line
-    if (util::CheckTrafficLightOnReferenceLine(reference_line_info,
-                                        traffic_light_overlap.object_id)) {
+    if (planning::util::CheckTrafficLightOnReferenceLine(
+        reference_line_info, traffic_light_overlap.object_id)) {
       continue;
     }
 
@@ -69,7 +70,7 @@ Stage::StageStatus TrafficLightUnprotectedRightTurnStageStop::Process(
     const double distance_adc_to_stop_line =
         traffic_light_overlap.start_s - adc_front_edge_s;
     auto signal_color =
-        util::GetSignal(traffic_light_overlap.object_id).color();
+        scenario::util::GetSignal(traffic_light_overlap.object_id).color();
     ADEBUG << "traffic_light_overlap_id[" << traffic_light_overlap.object_id
            << "] start_s[" << traffic_light_overlap.start_s
            << "] distance_adc_to_stop_line[" << distance_adc_to_stop_line
