+++ modules/planning/scenarios/util/util.h
@@ -18,32 +18,16 @@
 
 #include <string>
 
-#include "modules/common/vehicle_state/proto/vehicle_state.pb.h"
-#include "modules/planning/common/reference_line_info.h"
-#include "modules/routing/proto/routing.pb.h"
+#include "modules/perception/proto/traffic_light_detection.pb.h"
 
 namespace apollo {
 namespace planning {
+namespace scenario {
 namespace util {
 
-bool IsVehicleStateValid(const apollo::common::VehicleState& vehicle_state);
-
-bool IsDifferentRouting(const apollo::routing::RoutingResponse& first,
-                        const apollo::routing::RoutingResponse& second);
-
-double GetADCStopDeceleration(const double adc_front_edge_s,
-                              const double stop_line_s);
-
-bool CheckStopSignOnReferenceLine(
-    const ReferenceLineInfo& reference_line_info,
-    const std::string& stop_sign_overlap_id);
-
-bool CheckTrafficLightOnReferenceLine(
-    const ReferenceLineInfo& reference_line_info,
-    const std::string& traffic_light_overlap_id);
-
-bool CheckInsidePnCJunction(const ReferenceLineInfo& reference_line_info);
+perception::TrafficLight GetSignal(const std::string& signal_id);
 
 }  // namespace util
+}  // namespace scenario
 }  // namespace planning
 }  // namespace apollo
